feature: change to TestVariablesRecorded, and use it in ResultTransmissionEventHandler
