+++ models/classes/event/ResultTestVariablesTransmissionEvent.php
@@ -34,17 +34,21 @@ class ResultTestVariablesTransmissionEvent implements Event
     private $transmissionId;
     /** @var string */
     private $testUri;
+    /** @var bool  */
+    private bool $isManualScored;
 
     public function __construct(
         string $deliveryExecutionId,
         array $variables,
         string $transmissionId,
-        string $testUri = ''
+        string $testUri = '',
+        bool $isManualScored = null,
     ) {
         $this->deliveryExecutionId = $deliveryExecutionId;
         $this->variables = $variables;
         $this->transmissionId = $transmissionId;
         $this->testUri = $testUri;
+        $this->isManualScored = $isManualScored;
     }
 
     public function getName(): string
@@ -70,5 +74,10 @@ class ResultTestVariablesTransmissionEvent implements Event
     public function getTestUri(): string
     {
         return $this->testUri;
+    }
+
+    public function isManualScored(): bool
+    {
+        return $this->isManualScored;
     }
 }
